move comm files to dedicated folder
