+++ csrc/modules/parallel_llama4_decoder_module.h
@@ -3,7 +3,7 @@
 
 
 #include "../engine.h"
-#include "../comm_torch.h"
+#include "../comms/comm_torch.h"
 
 #include "parallel_multilinear_module.h"
 #include "parallel_llama4_attention_module.h"
