+++ csrc/parallel_gateupmoe_kernels.cu
@@ -1,5 +1,5 @@
 #include "./moeffn/gateupmoe.cuh"
-#include "comm_torch.h"
+#include "comms/comm_torch.h"
 
 #include <ATen/cuda/CUDAContext.h>
