+++ csrc/parallel_linear_kernels.cuh
@@ -2,7 +2,7 @@
 #define __MUILLM_PARALLEL_LINEAR_KERNELS_CUH__
 
 #include "engine.h"
-#include "comm_torch.h"
+#include "comms/comm_torch.h"
 #include "linear/linear.cuh" // for activation enum
 
 #include <optional>
